migrated weight_class_fvar to the OpenType profile
and moved its implementation and test-cases out of the fontwerk directory
